refactor(footer): add footer folder to components folder to improve files organization
