+++ src/components/dashboard/footer/footer.test.tsx
@@ -1,4 +1,4 @@
-import { DashboardFooter } from '@/components/dashboard/footer';
+import { DashboardFooter } from '@/components/dashboard/footer/footer';
 import { render, screen } from '@testing-library/react'; // TODO quando o arquivo test-utils.tsx estiver pronto, importar isso de lá
 
 describe('<DashboardFooter/>', () => {
