+++ src/components/dashboard/footer/footer.tsx
